Separate core logic into mcmcsample and mcmcpsample
